requirements.txt
fastapi>=0.94.0,<0.99.0
pydantic>=1.8.0,<2.0.0
uvicorn==0.27.1
sqlalchemy==1.4.25
psycopg2==2.9.6    #linux/mac user use: psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-jose
passlib
pydantic[email]
PyJWT
python-decouple
python-multipart
starlette>=0.27.0,<0.28.0
starlette-csrf